+++ api/src/routes/chrysalis/identity/diff/get.ts
@@ -1,14 +1,14 @@
 import * as identityLegacy from "@iota/identity-wasm-0.4/node";
 import * as identity from "@iota/identity-wasm/node";
-import { ServiceFactory } from "../../../factories/serviceFactory";
-import { IIdentityDiffHistoryBody } from "../../../models/api/IIdentityDiffHistoryBody";
-import { IConfiguration } from "../../../models/configuration/IConfiguration";
-import { CHRYSALIS } from "../../../models/db/protocolVersion";
-import { NetworkService } from "../../../services/networkService";
-import { IdentityHelper } from "../../../utils/identityHelper";
-import { ValidationHelper } from "../../../utils/validationHelper";
-import { IIdentityDiffHistoryRequest } from "./../../../models/api/IIdentityDiffHistoryRequest";
-import { IIdentityDiffHistoryResponse } from "./../../../models/api/IIdentityDiffHistoryResponse";
+import { ServiceFactory } from "../../../../factories/serviceFactory";
+import { IIdentityDiffHistoryBody } from "../../../../models/api/IIdentityDiffHistoryBody";
+import { IConfiguration } from "../../../../models/configuration/IConfiguration";
+import { CHRYSALIS } from "../../../../models/db/protocolVersion";
+import { NetworkService } from "../../../../services/networkService";
+import { IdentityHelper } from "../../../../utils/identityHelper";
+import { ValidationHelper } from "../../../../utils/validationHelper";
+import { IIdentityDiffHistoryRequest } from ".././../../../models/api/IIdentityDiffHistoryRequest";
+import { IIdentityDiffHistoryResponse } from ".././../../../models/api/IIdentityDiffHistoryResponse";
 
 /**
  * @param config The configuration.
